+++ RT09QT_HFT_2022231/RT09QT_HFT_2022231.Logic/CountyLogic.cs
@@ -2,6 +2,8 @@
 using RT09QT_HFT_2022231.Repository;
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Security.Cryptography;
 
 namespace RT09QT_HFT_2022231.Logic
 {
@@ -21,12 +23,22 @@ namespace RT09QT_HFT_2022231.Logic
 
         public void Delete(int id)
         {
-            this.repository.Delete(id);
+            var County = this.repository.Read(id);
+            if (County == null)
+            {
+                throw new ArgumentException("This county does not exist");
+            }
+                this.repository.Delete(id);
         }
 
         public County Read(int id)
         {
-            return this.repository.Read(id);
+            var County = this.repository.Read(id);
+            if (County == null)
+            {
+                throw new ArgumentException("This county does not exist");
+            }
+                return this.repository.Read(id);
         }
 
         public IEnumerable<County> ReadAll()
@@ -38,5 +50,20 @@ namespace RT09QT_HFT_2022231.Logic
         {
             this.repository.Update(county);
         }
+        //Folytköv
+        public int? GetTownCountPerCounty(int countyId)
+        {
+            return this.repository
+                .ReadAll()
+                .Where(t => t.CountyID == countyId).Count();
+        }
+
+        public int? GetInhabitantCountPerCounty()
+        {
+           
+            return from x in this.repository.ReadAll()
+        }
+
+        //Folytköv
     }
 }
